Code review fixes

+++ AnodaTurnTimer/Redux/AppActions.swift
@@ -32,10 +32,10 @@ struct RoundIsOutAction: Action {
     var beepValue: Int
 }
 
-struct RoundProgress: Action { var progress: Float }
+struct RoundProgress: Action { var progress: Double }
 struct RoundTimeInterval: Action { var timer: Int }
 
 struct RoundEndDate: Action { var endDate: Date? }
 struct RoundInitialAction: Action {
-    var progress: Float
+    var progress: Double
 }
